Give all the configuration options a unique prefix.
All of the options have been renamed to MixedCase, instead of
snake_case, and they all have the "PyUnit"-prefix.  This is true for all
options, except the options for project root detection support:
"ProjRootIndicators" and "ProjRootStopAtHomeDir", which are options that
I intent to reuse within my PEP8 and PyFlakes plugins.

See http://github.com/nvie/vim-pep8 and
http://github.com/nvie/vim-pyflakes for those.